backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-multipart
python-jose[cryptography]
